+++ src/windows.rs
@@ -1,98 +1,110 @@
 use crate::{DesktopEnv, Platform};
 
-	#[allow(unused)]
-	#[repr(C)]
-	enum ExtendedNameFormat {
-		NameUnknown, // Nothing
-		NameFullyQualifiedDN, // Nothing
-		NameSamCompatible, // Hostname Followed By Username
-		NameDisplay, // Full Name
-		NameUniqueId, // Nothing
-		NameCanonical, // Nothing
-		NameUserPrincipal, // Nothing
-		NameCanonicalEx, // Nothing
-		NameServicePrincipal, // Nothing
-		NameDnsDomain, // Nothing
-		NameGivenName, // Nothing
-		NameSurname, // Nothing
-	}
-	
-	#[allow(unused)]
-	#[repr(C)]
-	enum ComputerNameFormat {
-	  ComputerNameNetBIOS, // Same as GetComputerNameW
-	  ComputerNameDnsHostname, // Fancy Name
-	  ComputerNameDnsDomain, // Nothing
-	  ComputerNameDnsFullyQualified, // Fancy Name with, for example, .com
-	  ComputerNamePhysicalNetBIOS, // Same as GetComputerNameW
-	  ComputerNamePhysicalDnsHostname, // Same as GetComputerNameW
-	  ComputerNamePhysicalDnsDomain, // Nothing
-	  ComputerNamePhysicalDnsFullyQualified, // Fancy Name with, for example, .com
-	  ComputerNameMax
+use std::mem;
+
+#[allow(unused)]
+#[repr(C)]
+enum ExtendedNameFormat {
+    NameUnknown,          // Nothing
+    NameFullyQualifiedDN, // Nothing
+    NameSamCompatible,    // Hostname Followed By Username
+    NameDisplay,          // Full Name
+    NameUniqueId,         // Nothing
+    NameCanonical,        // Nothing
+    NameUserPrincipal,    // Nothing
+    NameCanonicalEx,      // Nothing
+    NameServicePrincipal, // Nothing
+    NameDnsDomain,        // Nothing
+    NameGivenName,        // Nothing
+    NameSurname,          // Nothing
 }
 
-	#[link(name = "Secur32")]
-	extern "system" {
-		fn GetUserNameExW(a: ExtendedNameFormat,
-            b: *mut u16,
-			c: *mut usize) -> u8;
-		fn GetUserNameW(a: *mut u16, b: *mut usize) -> i32;
-		fn GetComputerNameW(a: *mut u16, b: *mut usize) -> i32;
-		fn GetComputerNameExW(
-			a: ComputerNameFormat,
-            b: *mut u16,
-            c: *mut usize,
-		) -> i32;
-	}
+#[allow(unused)]
+#[repr(C)]
+enum ComputerNameFormat {
+    ComputerNameNetBIOS,             // Same as GetComputerNameW
+    ComputerNameDnsHostname,         // Fancy Name
+    ComputerNameDnsDomain,           // Nothing
+    ComputerNameDnsFullyQualified,   // Fancy Name with, for example, .com
+    ComputerNamePhysicalNetBIOS,     // Same as GetComputerNameW
+    ComputerNamePhysicalDnsHostname, // Same as GetComputerNameW
+    ComputerNamePhysicalDnsDomain,   // Nothing
+    ComputerNamePhysicalDnsFullyQualified, // Fancy Name with, for example, .com
+    ComputerNameMax,
+}
+
+#[link(name = "Secur32")]
+extern "system" {
+    fn GetUserNameExW(a: ExtendedNameFormat, b: *mut u16, c: *mut usize) -> u8;
+    fn GetUserNameW(a: *mut u16, b: *mut usize) -> i32;
+    fn GetComputerNameW(a: *mut u16, b: *mut usize) -> i32;
+    fn GetComputerNameExW(
+        a: ComputerNameFormat,
+        b: *mut u16,
+        c: *mut usize,
+    ) -> i32;
+}
 
 pub fn username() -> String {
-    let mut name = [0; 256];
-    let mut size = 256;
+    let mut name = mem::MaybeUninit::<[u16; 256]>::uninit();
+    let mut size = [256];
 
-    unsafe {
-        GetUserNameW(&mut name[0], &mut size);
-    }
+    let name = unsafe {
+        GetUserNameW(name.as_mut_ptr() as *mut _, size.as_mut_ptr());
+        name.assume_init()
+    };
 
-    String::from_utf16_lossy(if size == 0 { &[] } else { &name[..size - 1] })
+    String::from_utf16_lossy(if size[0] == 0 { &[] } else { &name[..size[0] - 1] })
 }
 
 #[inline(always)]
 pub fn realname() -> String {
-    let mut name = [0; 256];
-    let mut size = 256;
+    let mut name = mem::MaybeUninit::<[u16; 256]>::uninit();
+    let mut size = [256];
 
-    unsafe {
-        GetUserNameExW(ExtendedNameFormat::NameDisplay, &mut name[0], &mut size);
+    let name = unsafe {
+        GetUserNameExW(
+            ExtendedNameFormat::NameDisplay,
+            name.as_mut_ptr() as *mut _,
+            size.as_mut_ptr(),
+        );
+		name.assume_init()
+    };
+
+    if size[0] == 0 {
+        username()
+    } else {
+        String::from_utf16_lossy(&name[..size[0]])
     }
-
-	if size == 0 {
-		username()
-	} else {
-		String::from_utf16_lossy(&name[..size])
-	}
 }
 
 #[inline(always)]
 pub fn computer() -> String {
-    let mut name = [0; 256];
-    let mut size = 256;
+    let mut name = mem::MaybeUninit::<[u16; 256]>::uninit();
+    let mut size = [256];
 
-    unsafe {
-        GetComputerNameExW(ComputerNameFormat::ComputerNameDnsFullyQualified, &mut name[0], &mut size);
-    }
+    let name = unsafe {
+        GetComputerNameExW(
+            ComputerNameFormat::ComputerNameDnsFullyQualified,
+            name.as_mut_ptr() as *mut _,
+            size.as_mut_ptr(),
+        );
+		name.assume_init()
+    };
 
-    String::from_utf16_lossy(&name[..size])
+    String::from_utf16_lossy(&name[..size[0]])
 }
 
 pub fn hostname() -> String {
-    let mut name = [0; 256];
-    let mut size = 256;
+    let mut name = mem::MaybeUninit::<[u16; 256]>::uninit();
+    let mut size = [256];
 
-    unsafe {
-        GetComputerNameW(&mut name[0], &mut size);
-    }
+    let name = unsafe {
+        GetComputerNameW(name.as_mut_ptr() as *mut _, size.as_mut_ptr());
+		name.assume_init()
+    };
 
-    String::from_utf16_lossy(&name[..size])
+    String::from_utf16_lossy(&name[..size[0]])
 }
 
 pub fn os() -> String {